mathwizard_server/pyproject.toml
[project]
name = "mathwizard-server"
version = "0.1.0"
description = "Local FastAPI server for MathWizard using Gemma vision plus internal math tools."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "pydantic>=2.8.0",
  "python-multipart>=0.0.9",
  "sympy>=1.13.0",
  "uvicorn[standard]>=0.30.0",
  "Pillow>=10.4.0",
  "surya-ocr>=0.13.0",
]

[project.optional-dependencies]
transformers = [
  "accelerate>=0.34.0",
  "torch>=2.4.0",
  "transformers>=4.55.0",
  "torchvision>=0.19.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
